refactor InitialSettings
